+++ tests/framework/db/QueryBuilderTest.php
@@ -42,10 +42,10 @@ abstract class QueryBuilderTest extends DatabaseTestCase
     }
 
     /**
-     * @throws \Exception
-     * @return QueryBuilder
      * @param bool $reset
      * @param bool $open
+     * @return QueryBuilder
+     * @throws \Exception
      */
     protected function getQueryBuilder($reset = true, $open = false)
     {
@@ -696,13 +696,13 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 ],
             ],
             [
-                Schema::TYPE_STRING . ' CHECK (value LIKE "test%")',
-                $this->string()->check('value LIKE "test%"'),
+                Schema::TYPE_STRING . " CHECK (value LIKE 'test%')",
+                $this->string()->check("value LIKE 'test%'"),
                 [
-                    'mysql' => 'varchar(255) CHECK (value LIKE "test%")',
-                    'sqlite' => 'varchar(255) CHECK (value LIKE "test%")',
-                    'sqlsrv' => 'nvarchar(255) CHECK (value LIKE "test%")',
-                    'cubrid' => 'varchar(255) CHECK (value LIKE "test%")',
+                    'mysql' => "varchar(255) CHECK (value LIKE 'test%')",
+                    'sqlite' => "varchar(255) CHECK (value LIKE 'test%')",
+                    'sqlsrv' => "nvarchar(255) CHECK (value LIKE 'test%')",
+                    'cubrid' => "varchar(255) CHECK (value LIKE 'test%')",
                 ],
             ],
             [
@@ -726,13 +726,13 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 ],
             ],
             [
-                Schema::TYPE_STRING . '(32) CHECK (value LIKE "test%")',
-                $this->string(32)->check('value LIKE "test%"'),
+                Schema::TYPE_STRING . "(32) CHECK (value LIKE 'test%')",
+                $this->string(32)->check("value LIKE 'test%'"),
                 [
-                    'mysql' => 'varchar(32) CHECK (value LIKE "test%")',
-                    'sqlite' => 'varchar(32) CHECK (value LIKE "test%")',
-                    'sqlsrv' => 'nvarchar(32) CHECK (value LIKE "test%")',
-                    'cubrid' => 'varchar(32) CHECK (value LIKE "test%")',
+                    'mysql' => "varchar(32) CHECK (value LIKE 'test%')",
+                    'sqlite' => "varchar(32) CHECK (value LIKE 'test%')",
+                    'sqlsrv' => "nvarchar(32) CHECK (value LIKE 'test%')",
+                    'cubrid' => "varchar(32) CHECK (value LIKE 'test%')",
                 ],
             ],
             [
@@ -768,13 +768,13 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 ],
             ],
             [
-                Schema::TYPE_TEXT . ' CHECK (value LIKE "test%")',
-                $this->text()->check('value LIKE "test%"'),
+                Schema::TYPE_TEXT . " CHECK (value LIKE 'test%')",
+                $this->text()->check("value LIKE 'test%'"),
                 [
-                    'mysql' => 'text CHECK (value LIKE "test%")',
-                    'sqlite' => 'text CHECK (value LIKE "test%")',
-                    'sqlsrv' => 'nvarchar(max) CHECK (value LIKE "test%")',
-                    'cubrid' => 'varchar CHECK (value LIKE "test%")',
+                    'mysql' => "text CHECK (value LIKE 'test%')",
+                    'sqlite' => "text CHECK (value LIKE 'test%')",
+                    'sqlsrv' => "nvarchar(max) CHECK (value LIKE 'test%')",
+                    'cubrid' => "varchar CHECK (value LIKE 'test%')",
                 ],
             ],
             [
@@ -798,15 +798,15 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 ],
             ],
             [
-                Schema::TYPE_TEXT . ' CHECK (value LIKE "test%")',
-                $this->text()->check('value LIKE "test%"'),
+                Schema::TYPE_TEXT . " CHECK (value LIKE 'test%')",
+                $this->text()->check("value LIKE 'test%'"),
                 [
-                    'mysql' => 'text CHECK (value LIKE "test%")',
-                    'sqlite' => 'text CHECK (value LIKE "test%")',
-                    'sqlsrv' => 'nvarchar(max) CHECK (value LIKE "test%")',
-                    'cubrid' => 'varchar CHECK (value LIKE "test%")',
+                    'mysql' => "text CHECK (value LIKE 'test%')",
+                    'sqlite' => "text CHECK (value LIKE 'test%')",
+                    'sqlsrv' => "nvarchar(max) CHECK (value LIKE 'test%')",
+                    'cubrid' => "varchar CHECK (value LIKE 'test%')",
                 ],
-                Schema::TYPE_TEXT . ' CHECK (value LIKE "test%")',
+                Schema::TYPE_TEXT . " CHECK (value LIKE 'test%')",
             ],
             [
                 Schema::TYPE_TEXT . ' CHECK (value LIKE \'test%\')',
@@ -1087,10 +1087,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
 
     public function testCreateTableColumnTypes()
     {
-        if ($this->driverName === 'sqlsrv') {
-            $this->markTestSkipped('Should be fixed');
-        }
-
         $qb = $this->getQueryBuilder();
         if ($qb->db->getTableSchema('column_type_table', true) !== null) {
             $this->getConnection(false)->createCommand($qb->dropTable('column_type_table'))->execute();
@@ -1100,11 +1096,11 @@ abstract class QueryBuilderTest extends DatabaseTestCase
         foreach ($this->columnTypes() as $item) {
             list($column, $builder, $expected) = $item;
             if (!(strncmp($column, Schema::TYPE_PK, 2) === 0 ||
-                  strncmp($column, Schema::TYPE_UPK, 3) === 0 ||
-                  strncmp($column, Schema::TYPE_BIGPK, 5) === 0 ||
-                  strncmp($column, Schema::TYPE_UBIGPK, 6) === 0 ||
-                  strncmp(substr($column, -5), 'FIRST', 5) === 0
-                )) {
+                strncmp($column, Schema::TYPE_UPK, 3) === 0 ||
+                strncmp($column, Schema::TYPE_BIGPK, 5) === 0 ||
+                strncmp($column, Schema::TYPE_UBIGPK, 6) === 0 ||
+                strncmp(substr($column, -5), 'FIRST', 5) === 0
+            )) {
                 $columns['col' . ++$i] = str_replace('CHECK (value', 'CHECK ([[col' . $i . ']]', $column);
             }
         }
@@ -1155,7 +1151,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
             [['in', 'id', (new Query())->select('id')->from('users')->where(['active' => 1])], '[[id]] IN (SELECT [[id]] FROM [[users]] WHERE [[active]]=:qp0)', [':qp0' => 1]],
             [['not in', 'id', (new Query())->select('id')->from('users')->where(['active' => 1])], '[[id]] NOT IN (SELECT [[id]] FROM [[users]] WHERE [[active]]=:qp0)', [':qp0' => 1]],
 
-            [['in', 'id', 1],   '[[id]]=:qp0', [':qp0' => 1]],
+            [['in', 'id', 1], '[[id]]=:qp0', [':qp0' => 1]],
             [['in', 'id', [1]], '[[id]]=:qp0', [':qp0' => 1]],
             [['in', 'id', new TraversableObject([1])], '[[id]]=:qp0', [':qp0' => 1]],
             'composite in' => [
@@ -1300,13 +1296,10 @@ abstract class QueryBuilderTest extends DatabaseTestCase
      * @param array $condition
      * @param string $expected
      * @param array $expectedParams
+     * @throws \Exception
      */
     public function testBuildCondition($condition, $expected, $expectedParams)
     {
-        if ($this->driverName === 'sqlsrv') {
-            $this->markTestSkipped('Should be fixed');
-        }
-
         $query = (new Query())->where($condition);
         list($sql, $params) = $this->getQueryBuilder()->build($query);
         $this->assertEquals('SELECT *' . (empty($expected) ? '' : ' WHERE ' . $this->replaceQuotes($expected)), $sql);
@@ -1633,17 +1626,17 @@ abstract class QueryBuilderTest extends DatabaseTestCase
         $query = new Query();
         $secondQuery = new Query();
         $secondQuery->select('id')
-              ->from('TotalTotalExample t2')
-              ->where('w > 5');
+            ->from('TotalTotalExample t2')
+            ->where('w > 5');
         $thirdQuery = new Query();
         $thirdQuery->select('id')
-              ->from('TotalTotalExample t3')
-              ->where('w = 3');
+            ->from('TotalTotalExample t3')
+            ->where('w = 3');
         $query->select('id')
-              ->from('TotalExample t1')
-              ->where(['and', 'w > 0', 'x < 2'])
-              ->union($secondQuery)
-              ->union($thirdQuery, true);
+            ->from('TotalExample t1')
+            ->where(['and', 'w > 0', 'x < 2'])
+            ->union($secondQuery)
+            ->union($thirdQuery, true);
         list($actualQuerySql, $queryParams) = $this->getQueryBuilder()->build($query);
         $this->assertEquals($expectedQuerySql, $actualQuerySql);
         $this->assertEquals([], $queryParams);
@@ -2158,7 +2151,7 @@ abstract class QueryBuilderTest extends DatabaseTestCase
     }
 
     /**
-     * @depends testInitFixtures
+     * @depends      testInitFixtures
      * @dataProvider upsertProvider
      * @param string $table
      * @param array $insertColumns
@@ -2236,13 +2229,10 @@ abstract class QueryBuilderTest extends DatabaseTestCase
      * @param array $columns
      * @param array $value
      * @param string $expected
+     * @throws \Exception
      */
     public function testBatchInsert($table, $columns, $value, $expected)
     {
-        if ($this->driverName === 'sqlsrv') {
-            $this->markTestSkipped('Should be fixed');
-        }
-
         $queryBuilder = $this->getQueryBuilder();
 
         $sql = $queryBuilder->batchInsert($table, $columns, $value);