+++ tests/framework/db/SchemaTest.php
@@ -52,10 +52,6 @@ abstract class SchemaTest extends DatabaseTestCase
      */
     public function testGetTableNames($pdoAttributes)
     {
-        if ($this->driverName === 'sqlsrv') {
-            $this->markTestSkipped('Should be fixed');
-        }
-
         $connection = $this->getConnection();
         foreach ($pdoAttributes as $name => $value) {
             if ($name === PDO::ATTR_EMULATE_PREPARES && $connection->driverName === 'sqlsrv') {
@@ -68,14 +64,19 @@ abstract class SchemaTest extends DatabaseTestCase
         $schema = $connection->schema;
 
         $tables = $schema->getTableNames();
-        $this->assertTrue(\in_array('customer', $tables));
-        $this->assertTrue(\in_array('category', $tables));
-        $this->assertTrue(\in_array('item', $tables));
-        $this->assertTrue(\in_array('order', $tables));
-        $this->assertTrue(\in_array('order_item', $tables));
-        $this->assertTrue(\in_array('type', $tables));
-        $this->assertTrue(\in_array('animal', $tables));
-        $this->assertTrue(\in_array('animal_view', $tables));
+        if ($this->driverName === 'sqlsrv') {
+            $tables = array_map(static function ($item) {
+                return trim($item, '[]');
+            }, $tables);
+        }
+        $this->assertContains('customer', $tables);
+        $this->assertContains('category', $tables);
+        $this->assertContains('item', $tables);
+        $this->assertContains('order', $tables);
+        $this->assertContains('order_item', $tables);
+        $this->assertContains('type', $tables);
+        $this->assertContains('animal', $tables);
+        $this->assertContains('animal_view', $tables);
     }
 
     /**
@@ -197,7 +198,7 @@ abstract class SchemaTest extends DatabaseTestCase
 
     /**
      * @dataProvider tableSchemaCachePrefixesProvider
-     * @depends testSchemaCache
+     * @depends      testSchemaCache
      */
     public function testTableSchemaCacheWithTablePrefixes($tablePrefix, $tableName, $testTablePrefix, $testTableName)
     {
@@ -232,10 +233,6 @@ abstract class SchemaTest extends DatabaseTestCase
 
     public function testCompositeFk()
     {
-        if ($this->driverName === 'sqlsrv') {
-            $this->markTestSkipped('Should be fixed');
-        }
-
         /* @var $schema Schema */
         $schema = $this->getConnection()->schema;
 
@@ -532,7 +529,7 @@ abstract class SchemaTest extends DatabaseTestCase
             $this->assertSame($expected['scale'], $column->scale, "scale of column $name does not match.");
             if (\is_object($expected['defaultValue'])) {
                 $this->assertInternalType('object', $column->defaultValue, "defaultValue of column $name is expected to be an object but it is not.");
-                $this->assertEquals((string) $expected['defaultValue'], (string) $column->defaultValue, "defaultValue of column $name does not match.");
+                $this->assertEquals((string)$expected['defaultValue'], (string)$column->defaultValue, "defaultValue of column $name does not match.");
             } else {
                 $this->assertEquals($expected['defaultValue'], $column->defaultValue, "defaultValue of column $name does not match.");
             }
@@ -551,7 +548,7 @@ abstract class SchemaTest extends DatabaseTestCase
     public function testFindUniqueIndexes()
     {
         if ($this->driverName === 'sqlsrv') {
-            $this->markTestSkipped('Should be fixed');
+            $this->markTestSkipped('`\yii\db\mssql\Schema::findUniqueIndexes()` returns only unique constraints not unique indexes.');
         }
 
         $db = $this->getConnection();
@@ -803,7 +800,7 @@ abstract class SchemaTest extends DatabaseTestCase
         $newArray = [];
         foreach ($array as $value) {
             if ($value instanceof Constraint) {
-                $key = (array) $value;
+                $key = (array)$value;
                 unset($key['name'], $key['foreignSchemaName']);
                 foreach ($key as $keyName => $keyValue) {
                     if ($keyValue instanceof AnyCaseValue) {
@@ -843,7 +840,7 @@ abstract class SchemaTest extends DatabaseTestCase
             return;
         }
 
-        foreach (array_keys((array) $expectedConstraint) as $name) {
+        foreach (array_keys((array)$expectedConstraint) as $name) {
             if ($expectedConstraint->$name instanceof AnyValue) {
                 $actualConstraint->$name = $expectedConstraint->$name;
             } elseif ($expectedConstraint->$name instanceof AnyCaseValue) {