+++ framework/CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 Change Log
 2.0.23 under development
 ------------------------
 
+- Bug #17413, #17418: Fixed MSSQL tests (alexkart)
 - Bug #17413: Fixed MSSQL tests: `QueryBuilderTest::testGetColumnType()`, `QueryBuilderTest::testResetSequence()` (alexkart)
 - Bug #17420: Fixed loading of column default values for MSSQL (alexkart)